add and move extension classes
